+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## [Unreleased]
 
+### Added
+- Reformat code with spotless commit checkin handler
+
+### Changed
+- Run gradle task for the entire project in sync mode
+
 ## [1.0.4] - 2023-12-11
 
 ### Changed